// src/api/v1/bridge/types/bridge-status.type.ts
export enum BridgeStatus {
  CREATED = 'CREATED',
  APPROVAL_SIGNATURE_REQUIRED = 'APPROVAL_SIGNATURE_REQUIRED',
  APPROVAL_PENDING = 'APPROVAL_PENDING',
  BRIDGE_SIGNATURE_REQUIRED = 'BRIDGE_SIGNATURE_REQUIRED',
  BURN_SIGNATURE_REQUIRED = 'BURN_SIGNATURE_REQUIRED',
  MINT_SIGNATURE_REQUIRED = 'MINT_SIGNATURE_REQUIRED',
  SOURCE_PENDING = 'SOURCE_PENDING',
  ATTESTATION_PENDING = 'ATTESTATION_PENDING',
  STELLAR_SIGNATURE_REQUIRED = 'STELLAR_SIGNATURE_REQUIRED',
  DESTINATION_PENDING = 'DESTINATION_PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
}